requirements.txt
anthropic
beautifulsoup4
colorama==0.4.6
openai
playwright
PyYAML==6.0.2
Requests==2.32.3
tiktoken
lxml
python-dotenv
google-generativeai
streamlit
fpdf2